requirements.txt
boto3==1.37.18
botocore==1.37.18
h5py==3.13.0
loguru==0.7.3
matplotlib==3.10.1
mne==1.9.0
netCDF4==1.7.2
netCDF4==1.7.2
nibabel==5.3.2
nilearn==0.11.1
numpy==2.2.4
osfclient==0.0.5
pandas==2.2.3
Pillow==11.1.0
pytest==8.3.5
PyYAML==6.0.2
PyYAML==6.0.2
Requests==2.32.3
scikit_learn==1.6.1
scipy==1.15.2
seaborn==0.13.2
timm==1.0.15
torch==2.6.0
torchvision==0.21.0
tqdm==4.67.1
zenodo_get==1.6.1
